+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { CustomBackground } from './components/CustomBackground';
 import { SplashScreen } from './components/SplashScreen';
+import { LinksBar } from './components/LinksBar';
 import { PageWrapper } from './components/PageWrapper';
 import { About } from './components/pages/About';
 import { Contact } from './components/pages/Contact';
@@ -10,31 +11,34 @@ import { useBackgroundContext } from './backgrounds/BackgroundContext';
 export const defaultTransition = 'transition-all duration-500';
 
 export enum Screen {
+  Splash = 'Splash',
   About = 'About',
+  Projects = 'Projects',
   Contact = 'Contact',
 }
 
-const screensComponentMap: Record<Screen, JSX.Element> = {
+const screensComponentMap = {
+  [Screen.Splash]: <SplashScreen />,
   [Screen.About]: <About />,
   [Screen.Contact]: <Contact />,
 }
 
 export default function() {
-  const [screen, setScreen] = useState(Screen.About);
+  const [screen, setScreen] = useState<Screen>(Screen.Splash);
   const {zenMode} = useBackgroundContext()
 
   return (
     <div className='select-none bg-black text-pink-300 font-serif h-[100dvh] flex items-center justify-center flex-col'>
       <CustomBackground />
-      {!zenMode && <div className='bg-opacity-10 bg-black w-full h-full flex flex-col'>
-        <SplashScreen screen={screen} setScreen={setScreen} />
-        <div className='flex-1 relative  bg-opacity-70 bg-black rounded-3xl mt-5'>
-            {Object.entries(screensComponentMap).map(([screenName, component]) => (
-              <PageWrapper key={screenName} display={screen === screenName as Screen}>
-                {component}
-              </PageWrapper>
-            ))}
+      {!zenMode && <div className='absolute bg-opacity-10 bg-black w-full h-full flex items-center justify-center flex-col'>
+        <div className='flex items-center justify-center flex-1'>
+          {Object.entries(screensComponentMap).map(([screenName, component]) => (
+            <PageWrapper key={screenName} display={screen === screenName as Screen}>
+              {component}
+            </PageWrapper>
+          ))}
         </div>
+        <LinksBar screen={screen} setScreen={setScreen} />
       </div>}
         <FloatingIcons />
     </div>
